refactor: codebase cleanup and improvements
- Remove dead code: ViewedStateDismissed constant, ReviewThread.StartLine field
- Add typed DiffSide constant (DiffSideLeft/DiffSideRight) instead of raw strings
- Split model.go (766 lines) into model.go, view.go, commands.go
- Improve variable naming: c -> comment in client.go
- Replace custom containsString with strings.Contains in tests
- Add comments for intentionally ignored errors in parser.go
- Expand .gitignore with IDE, coverage, dist patterns
- Add errorlint and nolintlint to golangci-lint config
- Update CLAUDE.md to reflect new file structure
- Update README with lint command and requirements clarification
