+++ .golangci.yml
@@ -4,9 +4,11 @@ linters:
   default: none
   enable:
     - errcheck
+    - errorlint
     - govet
     - ineffassign
     - staticcheck
     - unused
     - gocritic
     - misspell
+    - nolintlint
